add namespace to demo scene script files move script files to demo scripts folder
