+++ sdkproject/Assets/Mapbox/Examples/Scripts/DragableDirectionWaypoint.cs.meta
