+++ sdkproject/Assets/Mapbox/Examples/Scripts/TrafficUvAnimator.cs.meta
